WIP; gatsby handle /groups splash pages;
this uses some recent work on the static site
to render splash pages for nextstrain groups
using the gatsby server and the new dataset
select interface.

Still to be addressed:
- General UI improvements such as rendering
custom markdown from each group and removing
possibly redundant info in the dataset/narrative
listings (e.g. contributor, the word dataset in the
case of using the dataset-select UI for narratives)

- getting access denied on s3 request for custom
group info (even in cases where it works on
nextstrain.org)

- getting internal error in handling getAvailable
 request like:
/charon/getAvailable?prefix=/groups/swiss/

- improve promise logic in componentDidMount

- more testing in general