Maintain a collection of group assets
Implements a new meta-source to represent the assets from all groups.
This is used for the API call /charon/getAvailable?prefix=/groups and
returns a listing of known datasets and narratives filtered to those
the user has permission to access.

As the collection of such assets involves many API calls, in order to
avoid excess server load this is stored as state and refreshed every 6
hours. This trades off staleness for reduced server load and quicker
API responses.

The implementation here is using a class interface inspired by our
`Source`s, however the methods differ. It is only instantiated from
within the `getAvailable` handler, rather than within our normal
flow (using `joinPartsIntoPrefix` and `splitPrefixIntoParts`). This
design choice adds the overhead of an extra "special case" but avoids
impacting all of our code paths which touch the two methods listed
above. Note that this code is only intended to be used for
/charon/getAvailable API calls at this stage.