parse /groups/:groupName markdown;
Copies auspice approach:
https://github.com/nextstrain/auspice/blob/master/src/util/parseMarkdown.js
but uses "sanitize-html" library:
https://github.com/apostrophecms/sanitize-html

A couple of config options from the auspice approach
didn't seem to have analogous options in the library I
use here. Also I don't add a hook to set the <a> tags
to open links in a new tab as is done in auspice. There's
probably a way to set this with the "sanitize-html" library
but it wasn't obvious to me and it wasn't working anyway
on nextstrain.org groups pages (which was using the auspice
code to parse markdown until recently for groups pages).