Implement panel components
As we expand the number of pages rendered by gatsby it becomes
more important to expose a small number of reusable components
which pages can use (see [1] for more).

Here we implement the concept of "panels" which are displayed in
a flexbox layout. Panels may contain custom content (we expose
`<Panel>` for this, or they may display a selection of cards
`<PanelWithCards`).

The exact breakpoints etc need further work, probably at a time
when the other components have been implemented.

Note that this commit intentionally includes content which should
be removed before release, either by subsequent commits or by
ammending this commit:
* /testing page
* wireframes around components

[1] https://github.com/nextstrain/nextstrain.org/issues/305