remove matchPath for /pathogens and /staging;
we were attempting to prefilter
listings on these pages according
to the * in e.g. /pathogens/* but
this was causing issues with linking
to auspice datasets from these pages.
This is not a high priority feature
and was not advertised yet so this
removes it (for now) to make sure
linking to auspice datasets from these
pages functions correctly.