Available staging data reflect bucket contents
This moves away from a hardcoded list of staging datasets (via the
manifest on s3://nextstrain-staging), which was out-of-date and
rarely updated. We now crawl the bucket so that all available datasets
are reported.

This introduces a significant speed penalty and the performance of the
server may not scale if many requests are made to list bucket contents.
Subsequent work should add a caching layer to `availableDatasets()`, or
similar.

The second tree options are now empty, which is an improvement on the
previous version which reported the options for the core (not staging)
source. This has the effect that the "Second Tree" dropdown is empty
when viewing staging datasets in Auspice.