Improve markdown parsing of URLs
Valid-looking URLs which we expect to be commonly used in markdown
(and, indeed, are being used in at least two group overviews) can
result in `new URL()` throwing a `TypeError`. Instead of failing
to parse the markdown, we instead allow the URLs but add the
suitable tags to the `<a>` element.

Closes https://github.com/nextstrain/nextstrain.org/issues/361