collect staging datasets instead of use manifest;
Currently the /staging page uses a /charon request
to source its dataset listing via the manifest. Here we
change to doing the following instead:
- Collect staging datasets with scripts/collect-datasets.js
  script and publish this list to s3 in a github action
- On /staging we request this listing from s3 to show a more
  frequently updated listing than with the manifest

This strategy is a comprimise between existing attempts:
(1)  using the manifest via a /charon request (current)
(2) crawling the s3 bucket via a /charon request (see https://github.com/nextstrain/nextstrain.org/pull/362)

in that it crawls the s3 bucket but on a regular interval
instead of with each request. Requests are made to a static
listing which is regenerated every 5 minutes (at no cost to the
nextstrain.org server). This should allow us to have caching
benefits that would be more complex to implement in (2) but
improves the freshness/accuracy of the listing compared to (1).

Some notes:
- currently the page is requesting the listing from https://staging.nextstrain.org/datasets_staging.json
  for testing purposes but we'll probably want this on
  data.nextstrain.org for production.
- we aren't getting narratives from the staging bucket.
  There aren't many there but we could do this if desired.
- I tacked the running of the collection script onto an existing
  GH action but there could be benefits to not coupling these udpates.
- I haven't implemented or configured any caching.
  There are many levels at which this can happen, and I'm
  not sure the degree we need for this case.