error behavior for /sars-cov-2
these changes allow you to go to
/ncov/*
/sars-cov-2/*
and see the sars-cov-2 page.
If you go to /ncov/* it's assumed
you want a dataset but if you land on
the sars-cov-2 gatsby page then that
dataset doesn't exist (otherwise you
would get routed to auspice) so we
display a banner letting you know that
the dataset doesn't exist but don't
wipe the url in case you want to subtly
change it and try again.

If you go to /sars-cov-2/*, any value of
* is assumed to be intended as a filter
you'd like to apply on the dataset listing.
E.g. /sars-cov-2/neherlab filters to
datasets matching the keyword neherlab.

This is just one part of a broader attempt
to behave consistently when a dataset
doesn't exist on nextstrain.org.
For more, see https://github.com/nextstrain/nextstrain.org/issues/355