error banner for missing dataset on more pages;
7788f4d showed a banner for non-existent
dataset urls for /ncov/* and /sars-cov-2/*.

This does the same thing for:
/influenza/*
/flu/*
/staging/*
/community/* (more to do here)

It also removes the auto-filtering of
dataset-listings according to the url
on any of the above pages.

Remaining questions/TODOs:
(1) As @jameshadfield put it:
Community URLs where the repo exists, but not the
dataset, are currently rendered in Auspice. These should be
moved to Gatsby, with an error banner atop a listing of the
datasets / narratives available for this source.
(2) Figure out what to do with pathogens/*.
Probably unlikely that people are visiting
/pathogens/pathogen-name but we could handle this
similarly to other pages if they are.
The tricky part is that most of these banners rely on
handling non-existent datasets under a certain path e.g.
/staging/*, by using Gatsby's `matchPath`. Meanwhile,
the core pathogens e.g. /zika, /ebola, etc.
should all probably end up on the /pathogens page if
someone tries a dataset that looks like one of those but
doesn't exist e.g. /zika/na. To achieve this the way we have
been handling things in gatsby-node we'd need to loop
over core pathogens something like this (some psuedo-code):
```
for p in [/zika, /ebola, ...]:
  createPage({
    path: p,
    matchPath: p/*,
    component: path.resolve("src/sections/pathogens-page.jsx")
  });
```
, which seems not very scalable/flexible.