fix community splash pages;
#368 implemented error behavior for various
pages including /community/*. In the case of
/community/* pages, this broke the ability to
visit a list of available datasets at a url like
/community/:user/:repo where user and
repo are valid [1].

Since we don't yet have /community/:user/:repo
splash pages in gatsby to list datasets, this
reverts to sending those to auspice to list datasets
for now.

A final note that when we do implement such pages,
they should be the ones to catch errors about invalid
user, repo, or dataset names, as opposed to the
top-level /community splash page. This is how we do
it for /groups as well.

[1] https://github.com/nextstrain/auspice/issues/1378