Align elements in each section horizontally
When sections are displayed side-by-side, it's nice to keep the
respective elements (cards, buttons) aligned.

To facilitate this I created a `<Section>` component for easier
editing of the styles. This should also make for nicer changes for
our future selves.

One regression here is the removal of `<ScrollableAnchor>` but these
can be added in if we want to keep this functionality.