[community] keep track of default branch
This commit is complementary to the previous commit. We now keep track
of the (underlying repo's) default branch, which allows us return
results in a form which mirrors the originating URL. Community URLs
(datasets or narratives) can be broken into three cases:

- "org/repo"
- "org/repo@branch" (branch is the default GitHub default branch)
- "org/repo@branch" (branch is not the default)

The getAvailable response returns results in the corresponding format.

Tests have been added to cover these.