Only add CORS headers for Gatsby dev server when the Express app is also in dev mode
Small thing, but good practice not to include in production things
needed only for dev.

Expands scope from single endpoint to all /charon/… endpoints, but this
is fine in dev.  We could even expand to the entire Express server, as
it's an artifact of Gatsby's dev server that they're separate and not
operating from the same origin.  This would make paths like /whoami
work.  I'll leave this larger change for another time, however.