Add Dataset.exists() method to replace datasetExists() function
With API and routing work, there's reason to do more existence checks in
places where a Dataset instance is available.  This also allows
subclasses of Dataset to implement appropriate alternate behaviour if
necessary.

The fetch() in sources.js is potentially caching (make-fetch-happen vs.
node-fetch in auspicePaths.js).  Although responses to HEAD requests
themselves aren't cached, a compatible GET which is already cached and
fresh may be used to generate a response to the HEAD here.  For that
reason, explicitly disable all caching with "no-store" mode.¹

In the future, we should consider opting back into caching to speed up
this check (and thus our routing).  Doing so will require further
consideration and likely modification of the cache-related response
headers our sources serve (or omit) as well as what other requests we
make which share the cache and can affect/be affected (hopefully for the
better!).  These reasons are why I'm not enabling it now.

The requests made here seemed to reliably trigger a TypeError (looking
up a property on an unexpectedly undefined value) in older versions of
make-fetch-happen, which is gone with an upgrade to the latest version
(included in this commit).  Good to upgrade anyway to catch other bug
fixes.

¹ https://github.com/npm/make-fetch-happen#opts-cache