charon: Treat 403s as 404s when fetching sidecar JSONs from upstream sources
S3 returns 403 Forbidden for a non-existent object if the principal has
s3:GetObject but not s3:ListBucket.¹  This is the case for many of our
public-access bucket policies.

If we need to distinguish 403s and 404s from non-S3 upstream sources in
the future, we can scope this check to only *.s3.amazonaws.com URLs.
For now, it's probably ~fine for current upstream sources that aren't S3
(github.com and /fetch/…), especially since non-sidecar dataset requests
return 404 if the upstream response is anything but a 200.

Avoids 500 Internal Server Errors from Charon endpoints when a
non-existent dataset sidecar is requested (e.g. root-sequence for many
of our datasets).

Note that errors thrown by this function, fetchJSON(), are only allowed
to percolate by the dataset sidecar codepath.  For the primary dataset
JSON, the v2 codepath doesn't use fetchJSON() and errors from the v1
codepath never see the light of day.

Adds useful detail to the 500 error message while I'm at it.

¹ https://docs.aws.amazon.com/AmazonS3/latest/API/API_HeadObject.html