Overhaul routing to be declarative
Declare ~all routes for nextstrain.org in the Express app and explicitly
define their endpoints.  This removes some imperative routing that
choose between Auspice and Gatsby and also makes explicit some routing
that was previously implicit.  The result is routing that's easier to
reason about and modify/extend later.

Modular endpoint functions (middleware and response handlers) are
defined for loading and validating sources, datasets, and narratives and
provide a foundation for adding more in the future API work.

The routing order is now:

  1. Declared routes backed by server-side endpoints, some of which serve
     the Auspice or Gatsby entrypoints
  2. Static filesystem-backed assets for Auspice mounted at /dist/…
  3. Static filesystem-backed pages and assets for Gatsby mounted at /…
  4. Gatsby 404 page (with status 404), which may invoke Gatsby's
     client-side routing

Gatsby's client-side routing is thus the final backstop when a request
doesn't match any route, as we serve our Gatsby 404 page.  If Gatsby's
client-side routes match the request, then Gatsby will display the
matching page; otherwise the 404 page is shown.  Standard Gatsby pages
should be served directly by #3 above, so client-side routing from the
404 page in #4 should only happen for pages which act as custom 404
pages, e.g. when a group dataset doesn't exist and we show the
individual group page itself with an error instead of the generic 404
page.