Fix mixed-up Community narrative routes
In "Overhaul routing to be declarative" (d5ef4b7), I mistakenly setup
routes for

  /community/:user/:repo/narratives

instead of

  /community/narratives/:user/:repo

Oops!  (Notably, the Groups routes are /groups/:groupName/narratives
instead of /groups/narratives/:groupName because the former was and is
more natural/intuitive to me.)

I realized when fixing the routes that Community narratives have the
same property as Community datasets in that they can optionally be
"path-less" (based only on repo name), so I updated the comment and
logic to apply to narratives as well.  Note, however, that our current
repo-level Community page in Gatsby assumes every missing thing is a
dataset and not a narrative.  We'll want to fix that in a subsequent
change.

This mistake was easy to make with imperative routing because you have
to "be the computer" and evaluate the code (e.g. splitPrefixIntoParts())
in your head to know what routes exist.  Declarative routing means you
can just look at the list of routes.