Fix setSource() call for /fetch routes
The argsExtractor was returning a string instead of an array of args,
meaning that UrlDefinedSource(...argsExtractor(req)) produced a bunch of
single-character arguments instead of a single string argument.

I didn't catch this when writing it because the setSource() here isn't
actually used at the moment; it's anticipatory only and parallels the
routes for other sources.  I only caught it in development when I added
additional endpoints that starting using the source context object.
This sort of bug is exactly why Typescript would be nice to adopt.