negotiate: Support middleware handlers in contentTypesProvided()
Each media type can now be followed by a stack of handlers instead of
just one, like Express route handlers (but without the unnecessary
array-flattening behaviour).

All but the last handler in the stack should call "return next()" or
"await next()" in order to chain to the next handler in the stack
(unless that is not desired).