Move static handlers for Auspice and Gatsby into endpoints/static
Better separation of concerns here.  Note that instead of using the
"fallthrough" option to expressStaticGzip (which passes it thru to
express.static, which passes it thru to serve-static), I stack an
"always 404" handler after the static handler.  Otherwise, the
fallthrough behaviour would be non-obvious and baffling to anyone
reading app.js.  Falling through is part of the routing, which belongs
solely in app.js, while how to serve the assets is part of the endpoint.