Organize most endpoints under a single object
Starts to reduce function proliferation, and it seems clearer to use
"endpoints" as the housing object instead of extending the pattern of
"charon" and "users" to "static" (can't actually use that name) and
"sources" (potentially confusing with ./src/sources.js).

Source-related endpoints are still destructured into individual
variables because they're used in so many routes and repeating
"endpoints.sources." all over the place would be harder to read.