tests: Run all tests in test/smoke-test/
Simplifies the test file discovery for both kinds of tests by defining a
global glob match on filename and then restricting discovery to each
respective test directory when invoking Jest.  The defaults for Jest are
overly complex for our needs.

Previously the filenames we passed to Jest were treated as regexp
patterns for filtering its global search for test files, not a simple
list of files.