Read-only RESTful API endpoints for datasets and narratives
Extends GET endpoints for existing dataset and narrative routes to
provide the following media types via content negotiation:

    application/vnd.nextstrain.dataset.main+json
      (also aliased as just application/json)
    application/vnd.nextstrain.dataset.root-sequence+json
    application/vnd.nextstrain.dataset.tip-frequencies+json

    text/vnd.nextstrain.narrative+markdown
      (also aliased as just text/markdown)

A new documentation page covers how to use the API.  I wrote the page in
rST instead of Markdown because it's more expressive, and my intention
is for the page to be incorporated into docs.nextstrain.org eventually.
Not sure yet if this repo eventually becomes another RTD subproject or
if we do cross-repo sourcing of the pages instead.

Documentation includes prospective support for the PUT and DELETE
methods that are not included in this commit but will come later.