Hone heuristic for "is this request browser-like?" in error handling
Essentially flips the default/fallback value to be "no" instead of "yes"
by checking for explicit inclusion of text/html instead of if text/html
is merely acceptable.  This is because sending Accept: */* by default is
a very common practice of many clients (browser and non-browser).  ~All
browsers seem to declare text/html explicitly, while other clients
rarely do, so this works better in practice.

Updates a test fetch() to be more browser-like when testing our HTML
error handling.