charon: Fix duplicated repo name in getAvailable response for /community
Resolves #121.

The replace in CommunitySource.availableDatasets() that always included
a trailing slash failed to consider "top-level" community datasets (e.g.
repoName.json instead of repoName_datasetName.json).

Although availableNarratives() was doing it right, switch it to using
the same removal function for consistency.

I considered changing the replace pattern to a regexp anchored at the
end with (/|$), but this requires escaping/quoting regexp metachars in
repoName and it turns out there is no handy function to do that.  It'd
be simple in other regexp flavors with interpolation and \Q and \E, but
having to add a quote function plus having to construct the regexp with
string concatenation made for something that seemed less clear than
doing startsWith() + slice() in a method.