Don't check existence of datasets and narratives for /fetch/…
This fixes a key use case of /fetch where AWS S3 signed URLs are given.

These signed URLs dictate the permitted S3 action, and the same URL
can't support both a HEAD (headObject) and GET (getObject) request.
Prior to the merge and deploy of #443 this morning, the /fetch endpoints
by design didn't guard Auspice's entrypoint with a HEAD request for the
dataset to check existence.  This inadvertently changed in "Refactor
endpoints for GETing a dataset/narrative to make room for additional
media types" (86190b7).

Now once again for /fetch we'll always serve up Auspice, which will make
a request to Charon, which may return an error if the upstream request
fails, and an error from Charon will trigger Auspice's error page.
That's unlike everywhere else where we try to avoid Auspice's error
page, but it's ok.

Resolves #461.