sources: Define Groups as JSON instead of code
A step towards new Groups only requiring a change to this data file to
provision them (once a multi-tenant bucket is used).

One day we'll probably store this Groups information in a database table
(probably when we need to move Groups membership information out of
Cognito), but a simple JSON file is perfectly fine for now.  A data
model class acts as the access layer on top of the data store, so it's
easy to swap out the store (i.e. JSON file) in the future.

Resolves #183.