sources: Stop using source "names" separate from the classes themselves
The new unified GroupSource which is parameterized by group name allows
us to switch from an indefinitely growing list of source classes to just
five and get rid of these secondary string identifiers.  This also makes
the isGroup() method unnecessary and so it is removed.  What was the
GroupSource function briefly is now a class so that we can properly
introspect it, e.g. with instanceof.  A parameterized class wasn't
possible before removing the static _name property, so public vs.
private had to be separate still and papered over with the previous impl
as a plain function.

Several places which were relying on the assumption that source.name
equals Group name for Nextstrain Groups no longer rely on that
assumption, which plugs some abstraction leaks.  All authz decisions are
still made on a Source object, for now, but we could extend this to
Group objects too in the future if it makes sense.

The CoreSource and CoreStagingSource still use a .name object property
(but not a ._name class property) to support code shared between them
which still needs the strings "core" and "staging".  The implementation
of Dataset.resolve() moves into CoreSource (from Source) so it can keep
relying on .name.  It only ever had a purpose for CoreSource and
CoreStagingSource anyway and was put in Source to avoid a custom
CoreDataset (which this commits adds now; it's not so bad).

GroupSources now always use signed S3 URLs for both public and private
groups, instead of public groups using public S3 URLs for GET/HEAD.
This is simpler and the public S3 URLs provided no real benefit.  It
also aligns with the future migration of all groups into a single
multi-tenant S3 bucket, where public S3 URLs won't be an option.

Resolves #76.