sources: Customize how classes and objects stringify
This makes it possible to throw them into logging statements and get
something reasonable out.  By default, classes stringify to their source
code (!) and objects to "[object Object]", neither of which are very
useful.

In the future, we could override these in subclasses to add further
details, e.g. the GitHub repo for CommunitySource.