sources: Do away with the fiction of a separate S3Source
While it wasn't specific to Groups sources when it was first added, it
was only ever used for Groups (despite other sources also being backed
by S3) and so over time accumulated Groups-specific changes.  For
simpler code structure, fold S3Source into GroupSource.  We can extract
a generic S3Source back out if it makes sense to in the future.