sources: Avoid listing bucket objects just for the group logo and overview
Listing objects in a bucket can be slow, especially as the bucket grows.
Instead, construct URLs with Source.urlFor() and do two normal
fetch()-es for group-logo.png and group-overview.md.

An upshot is that this removes one of the outlying codepaths that didn't
go through urlFor() and thus didn't get centralized path handling, which
will matter more with the coming switch to a multi-tenant bucket.

As a bonus of using fetch(), we get automatic decompression instead of
having to handle it ourselves and caching with revalidation using
conditional requests.

The logo URL returned to clients now uses the data: scheme instead of
passing through an S3 signed URL.  I figured the server has to do a
fetch anyway to determine existence and might as well save the client a
fetch.