sources: Improve parsing of group-overview.md frontmatter
Make it more robust with better validation checks and more resilient
with resetting of specific bad fields instead of total failure.  Cleans
up the structure of the code as well, e.g. so defaults are shared.

The website URL's protocol is checked against a safe set instead of an
easily-bypassed substring check.  The type of showDatasets and
showNarratives is checked even if the value is falsey (but not false).
Title is no longer required since it is not actually necessary.