Track and manage a couple critical AWS IAM policies in this repo
These policies are closely coupled to the code in this repo and used for
the nextstrain.org production server as well as development/testing
servers.  I do not consider these policies to be sensitive as the
information in them is inferable from the existing codebase.  The
current state of the policies is committed here; an update will follow
in a separate commit.

While this sort of thing is the bread-and-butter of tools like Terraform
and Pulumi, and we should probably eventually adopt one of those, those
tools are much more involved to get started with and would be
unnecessary overhead and complexity right now.  A basic sync program is
still better than manual updates (our current method) and helps us more
plausibly use this repo as the system of record for what the policies
should be.

The sync program is written in Python because the boto3 API for IAM is
_much_ nicer to use than the AWS SDK for JS API.  boto3 is the only
dependency not in the standard library.  I expect the operators of the
program to be well-versed enough in Python to be able to run it in an
appropriate environment.

Sync is still manually operated, not run automatically on push.  In the
future, we might want to choose to allow CI to manage these policies
(and just these policies).