groups: Store mapping from generic roles → authz roles on the Group
The immediate impetus for this is a new Group provisioning script which
needs to a) enumerate all the authz roles to create Cognito groups as
well as b) convert from generic → fully-qualified names to make inputs
nicer for the operator.

More generally, centralizing the mapping will help avoid a proliferation
of hardcoded construction of these authz roles.  It also makes it easier
to add custom roles or adjust the mapping on a special-case basis.