Automate provisioning of a new Group
Program is idempotent given the same input and cooperates with existing
Cognito state, thus making it suitable for adding members to existing
Groups too.

Code in this can potentially be extracted and re-used for future Group
management features.  I've tried to keep the functions that take actions
in Cognito tightly-scoped and separate from the interface and reporting
code.

Uses v3 of the AWS SDK for JS to take advantage of improved concurrency
and async/await support.