provision-group: Add a user "mustExists" flag for avoiding accidental creation
This is particularly useful as a safety check when adjusting membership
roles or otherwise adjusting existing user information.  I'll use this
flag to help provision new Cognito groups and memberships for existing
early adopter Groups, where all users should already exist.

The flag is per-user in the members file so that members who are
existing users can be mixed with users who aren't in the case of
manually-constructed members files.