groups: Move byline default to overview content default
The default doesn't make sense as a "byline".  It makes more sense as a
stand-in for a custom overview of the group.  If we want to de-emphasize
the default with different styling (do we?), we should condition on that
instead of abusing byline.  It's also easy to imagine a future default
byline that's actually a byline by displaying the group members.