Adjust layout of multi-tenant groups bucket
The bucket now stores datasets and narratives separately under a fixed
prefix within each group's prefix so the user-provided names/paths of
the datasets and narratives can't conflict with control objects we store
at the root of the group's prefix.

Besides making security easier to reason about here with the
user-provided names, this will also, in time, allow us to eliminate a
few special-cased exclusions (like excluding group-overview.md from a
list of *.md narratives).

As part of deploying this, we'll need to perform a one-time manual
layout switch behind the scenes using scripts/migrate-groups-layout.
Though migrate-groups-layout is intended to be run only twice:

  1. Immediately before deploy of the layout change.
  2. Immediately after deploy of the layout change, with --delete-after-copy.

…it is designed to be safe to run repeatedly.  However, as it is a
single-use program, it's not designed to live past those two primary
invocations and we should delete it after it's served its purpose.