FilterSelect: Use virtualized menu list
Add a VirtualizedMenuList component to replace the MenuList component of
the react-select package.

This was added to replace the `react-virtualized-select` package that is
no longer maintained and incompatible with the latest version of
`react-select`.

This component is slightly different than the one in Auspice because
it requires an `AutoSizer` to make a responsive list that does not have
a fixed width.