authn: Share our session cookie across nextstrain.org and subdomains
By setting the cookie's domain, it changes from a host-only cookie (the
default, limited to just the host that sets it, e.g. nextstrain.org) to
a domain-wide cookie (including all levels of subdomains).¹

This will allow sessions from nextstrain.org to be portable to
next.nextstrain.org and dev.nextstrain.org, and vice versa.  These
instances already share session storage (a Redis instance) and the
secret used to secure sessions (SESSION_SECRET).  Using a domain-wide
cookie was the last missing piece.

Note the security implications of this change, added as a comment for
maximum visibility.

Also note the cookie name change and transparent migrating of existing
sessions, as explained in comments, which is necessitated by the
host-only → domain-wide switch to avoid oodles of confusion.

Related to <https://github.com/nextstrain/nextstrain.org/issues/485> and
<https://github.com/nextstrain/nextstrain.org/issues/492>.

¹ I'd been under the impression that a leading dot on the cookie domain,
  e.g. .nextstrain.org, was necessary for this behaviour.  That was the
  case for ages with RFC2109 but has long since been outdated by
  RFC6265.