Automatically redirect to canary deployment for "canary"-flagged users
Applies only to session-authenticated requests, which is expected to be
interactive use by a human in a browser (who made a choice to prefer to
use the canary).  There are other ways for an API client to opt-into
canary usage (e.g.  setting NEXTSTRAIN_DOT_ORG for the Nextstrain CLI).

Resolves <https://github.com/nextstrain/nextstrain.org/issues/492>.