authn: Correct comment about Cognito client ids
It's true that the dev client id used to name https://dev.nextstrain.org
as a valid origin in our Cognito client config.  However, that never
mattered because dev.nextstrain.org always used the *prod* client id
since it runs in a production env (e.g. on Heroku) too.  So the
commentary here was technically correct, but misleading even when I
wrote it.  In the meantime, we've since sorted out our desired behaviour
with the Cognito client configs (and related Heroku env vars), but the
comment wasn't updated to match.