authn: Preserve original Cognito groups on user object
This will make potential future user object updates in deserializeUser()
much easier because we won't have to figure out which old user
properties contain the "unparsed" or "original" array of Cognito groups
to re-parse.

Ultimately, it will still be better to move to storing the original id
and refresh tokens in the session instead of the user object.  The
deserializeUser() step would then reparse the id token (after
potentially refreshing it) every time and not have to worry about
upgrading a derived, persistent object.  I'm not sure when we'll get
around to that, however, so for now this bit of low-effort
future-proofing seems likely to be helpful.

Related to <https://github.com/nextstrain/nextstrain.org/issues/81>.