migrate-group: Actually remove IAM group members
All members of the IAM group must be removed before the IAM group can be
deleted later on.  This was missed in dev/testing because all the test
groups I used only had the typical "nextstrain-${groupname}" IAM user,
which gets deleted (and thus is no longer an IAM group member).
While this code expects that some of our IAM groups for Groups have
additional IAM users in them (and won't delete those users), it forgot
to remove their membership.