Community dataset + narrative listing
This implements a listing of (known) community pages via a hardcoded
YAML file. This is the first implementation, and there are plenty
of improvements to be made when we systematically address how assets
are collected and displayed.

I am aware of plenty more datasets and narratives than these, however
we want to make this an opt-in listing of resources.

Closes #486