Add helper script to parse community data
We wanted to extract metadata (specifically the maintainer / author)
from community datasets / narratives & it was easier to write a small
script than examine every page by hand!

See added comments (in code + yaml file) for how to run. Eventually
this should be automated as the current situation is likely to fall
out of date, but this is a good first step.