Request my database  from ROK side
Dear Nextstrain team,
I sent a comment on behalf of Professor Jin-Won Song from the Republic of Korea (ROK).
I am a researcher at Korea University in the ROK. 
I am pleased to inform you that our analysis for interested viruses using nextstrain workflow (system) is finished. 
We hope that our dataset is accessible via a front web page (Nextstrain community page).
We want to replace the database of the hantavirus, influenzavirus, SFTSV that was posted previously.
I provided the URL submitted to Github.
I would appreciate it if you could post our dataset on the front page (or community page) in the Nextstrain web page.
Please let me know the due date of the process for posting our datasets and give me a reply if you have any questions.