Actually respond with 401 Unauthorized as intended under certain conditions
The intent was for AuthzDenied errors to be translated to 401
Unauthorized and 403 Forbidden responses depending on the context, but
this distinction was accidentally lost in "Handle authz failures in the
app-wide error handler" (fe2e5bd).  What were throws became assignments
to "err" without corresponding changes to keep the control flow the
same.  The result was that all AuthzDenied errors became 403s.