authn: Switch to storing user tokens in sessions instead of the user object
Validates the stored tokens and constructs a user object on each
request, with automatic behind-the-scenes token renewal (and thus user
info refresh) as necessary.  This is identical to the Nextstrain CLI
token handling.  We can control how often the server refreshes info from
Cognito by tuning the token lifetimes in the Cognito client config for
the server.  This change also lets us implement "stale before"
timestamps for users when we knowingly change their info in Cognito and
automatically refresh based on those so that we can keep the cached data
fresh.

Existing sessions are left as-is for now.  They can't be upgraded
automatically because the tokens used to establish them at initial login
are long gone.  We'll need to decide what to do with them (e.g. force
all existing sessions to login again).

Partially resolves <https://github.com/nextstrain/nextstrain.org/issues/81>.