Weaken upstream ETag if we decompress the upstream response body in transit
A strong ETag is used for exact equality (byte-for-byte), while weak
ETags mean only equivalent response bodies (such as compressed
versions).  Compressed versions can have separate strong ETags, but
generation and management of those is beyond the scope of this code.

This small bug would only be noticed by clients which cache responses
based on ETag and also issue requests for the same URLs where sometimes
they Accept-Encoding: gzip and sometimes they don't.  And it would only
be noted by a client comparing two responses (compressed vs.
uncompressed) with the (buggy) identical ETag.  I noticed it when
investigating possible use of and support for If-None-Match.